# perf(helm): load in-memory files