## Description

Instead of copying files to a temporary folder, we can pass virutal files to helm, which improves performance.

```bash
benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/aquasecurity/trivy/pkg/iac/scanners/helm/test
                    │   old.txt    │               new.txt               │
                    │    sec/op    │   sec/op     vs base                │
_RenderChartFiles-8   2138.1µ ± 2%   978.3µ ± 3%  -54.24% (p=0.000 n=10)

                    │   old.txt    │               new.txt               │
                    │     B/op     │     B/op      vs base               │
_RenderChartFiles-8   570.4Ki ± 0%   532.8Ki ± 0%  -6.59% (p=0.000 n=10)

                    │   old.txt   │              new.txt               │
                    │  allocs/op  │  allocs/op   vs base               │
_RenderChartFiles-8   7.714k ± 0%   7.476k ± 0%  -3.09% (p=0.000 n=10)
```

## Checklist
- [x] I've read the [guidelines for contributing](https://aquasecurity.github.io/trivy/latest/community/contribute/pr/) to this repository.
- [x] I've followed the [conventions](https://aquasecurity.github.io/trivy/latest/community/contribute/pr/#title) in the PR title.
- [x] I've added tests that prove my fix is effective or that my feature works.
- [ ] I've updated the [documentation](https://github.com/aquasecurity/trivy/blob/main/docs) with the relevant information (if needed).
- [ ] I've added usage information (if the PR introduces new options)
- [ ] I've included a "before" and "after" example to the description (if the PR is a user interface change).