@simar7 2024-03-26
@nikpivkin just curious if there's any special handling that we need to perform with symlinks, or would this work with that as well?