@nikpivkin 2024-03-26
@simar7 As far as I know, Trivy [skips](https://github.com/aquasecurity/trivy/blob/main/pkg/fanal/walker/fs.go#L61) non-regular files, so they don't get into the virtual file system that post analysers work with.